+++ Scripts/WoW Classic Game Data API/Realm API/Endpoint_WowClassic_GameData_Realm.cs
@@ -1,0 +1,118 @@
+// ╔════════════════════════════════════╗
+// ║ This file has been auto-generated. ║
+// ╚════════════════════════════════════╝
+
+using System;
+using System.Collections;
+using ZeShmouttsAssets.BlizzardAPI.JSON;
+
+namespace ZeShmouttsAssets.BlizzardAPI
+{
+	/// <summary>
+	/// Interface for working with the Blizzard API inside Unity.
+	/// </summary>
+	public static partial class BlizzardAPI
+	{
+		/// <summary>
+		/// API endpoints related to World of Warcraft Classic game data (items, spells, etc.).
+		/// Reference : https://develop.battle.net/documentation/world-of-warcraft-classic/game-data-apis
+		/// </summary>
+		public static partial class WowClassicGameData
+		{
+			internal const string apiPath_Realm = basePath_Wow_gameData + "/realm/{0}";
+
+			/// <summary>
+			/// Coroutine that retrieves a single realm by slug or ID.
+			/// </summary>
+			/// <param name="realmSlug">The slug of the realm.</param>
+			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetRealm(string realmSlug, Action<Json_WowClassic_Realm> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = string.Format(apiPath_Realm, realmSlug);
+				yield return SendRequest(region, namespaceClassicDynamic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+			/// <summary>
+			/// Coroutine that retrieves a single realm by slug or ID.
+			/// </summary>
+			/// <param name="realmId">The ID of the realm.</param>
+			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetRealm(int realmId, Action<Json_WowClassic_Realm> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = string.Format(apiPath_Realm, realmId);
+				yield return SendRequest(region, namespaceClassicDynamic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+			/// <summary>
+			/// Coroutine that retrieves a single realm by slug or ID, as a raw JSON string.
+			/// </summary>
+			/// <param name="realmSlug">The slug of the realm.</param>
+			/// <param name="action_Result">Action to execute with the raw JSON string.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetRealmRaw(string realmSlug, Action<string> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = string.Format(apiPath_Realm, realmSlug);
+				yield return SendRequest(region, namespaceClassicDynamic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+			/// <summary>
+			/// Coroutine that retrieves a single realm by slug or ID, as a raw JSON string.
+			/// </summary>
+			/// <param name="realmId">The ID of the realm.</param>
+			/// <param name="action_Result">Action to execute with the raw JSON string.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetRealmRaw(int realmId, Action<string> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = string.Format(apiPath_Realm, realmId);
+				yield return SendRequest(region, namespaceClassicDynamic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+		}
+	}
+}
+
+namespace ZeShmouttsAssets.BlizzardAPI.JSON
+{
+	/// <summary>
+	/// JSON structure for World of Warcraft (Classic), representing a realm.
+	/// </summary>
+	[Serializable]
+	public class Json_WowClassic_Realm : Object_JSON
+	{
+		// {{JSON_START}}
+		public LinkStruct _links;
+
+		public int id;
+		public RefNameIdStruct region;
+		public HRefStruct connected_realm;
+		public LocalizedString name;
+		public LocalizedString category;
+		public string locale;
+		public string timezone;
+
+		[Serializable]
+		public struct RealmType
+		{
+			public string type;
+			public LocalizedString name;
+		}
+		public RealmType type;
+		public bool is_tournament;
+		public string slug;
+		// {{JSON_END}}
+	}
+}